diary id를 입력받는 곳을 long으로 변환

+++ src/main/java/com/fooddiary/api/repository/diary/DiaryRepository.java
@@ -13,10 +13,10 @@ import java.util.Optional;
 
 
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
 
     @Query("select d from Diary d join fetch d.images i where d.id = :id order by i.updateAt desc")
-    Optional<Diary> findDiaryAndImagesById(@Param("id") int diaryId);
+    Optional<Diary> findDiaryAndImagesById(@Param("id") long diaryId);
 
     @Query("select d from Diary d inner join fetch d.images i where d.createTime between :startDate and :endDate and d.user.id = :userId order by d.createTime asc, i.updateAt desc")
     List<Diary> findByYearAndMonth(@Param("startDate") LocalDateTime startDate, @Param("endDate") LocalDateTime endDate, @Param("userId") int userId);
@@ -32,10 +32,10 @@ public interface DiaryRepository extends JpaRepository<Diary, Integer> {
                                      @Param("endDate") LocalDateTime endDate, @Param("userId") int userId);
 
     @Query("select count(d.id) from Diary d inner join d.images where d.id = :id")
-    int getDiaryImagesCount(@Param("id") int id);
+    int getDiaryImagesCount(@Param("id") long id);
 
     @Query("select d from Diary d left join fetch d.images i where d.user.id = :userId and d.id > :id order by d.id asc")
     List<Diary> findByUserIdAndLimit(@Param("userId") int userId, @Param("id") long id, Pageable pageable);
 
-    Diary findByUserIdAndId(@Param("userId") int userId, @Param("id") int id);
+    Diary findByUserIdAndId(@Param("userId") int userId, @Param("id") long id);
 }
